backend/requirements.txt
fastapi==0.110.2
uvicorn==0.29.0
pymongo==4.6.3
python-dotenv==1.0.1
passlib[bcrypt]==1.7.4
python-jose==3.3.0
sentence-transformers==2.7.0
keybert==0.8.3
scikit-learn==1.4.1.post1
beautifulsoup4==4.12.3
requests==2.32.3
scikit-learn==1.3.2